refactor(songs): remove old `Song` model
