+++ lib/navigation.dart
@@ -5,8 +5,8 @@ import 'package:musbx/drone/drone_page.dart';
 import 'package:musbx/metronome/metronome_page.dart';
 import 'package:musbx/songs/library_page/library_page.dart';
 import 'package:musbx/songs/player/song.dart';
-import 'package:musbx/songs/player/song_source.dart';
 import 'package:musbx/songs/player/songs.dart';
+import 'package:musbx/songs/player/source.dart';
 import 'package:musbx/songs/song_page/song_page.dart';
 import 'package:musbx/tuner/tuner_page.dart';
 import 'package:musbx/utils/persistent_value.dart';
@@ -80,7 +80,7 @@ class Navigation {
                     final String id = state.pathParameters["id"]!;
 
                     // Begin loading song
-                    final SongNew song = Songs.history.map.values
+                    final Song song = Songs.history.map.values
                         .firstWhere((song) => song.id == id);
 
                     return FutureBuilder(
