+++ lib/songs/analyzer/chord_identification_process.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:musbx/model/chord.dart';
 import 'package:musbx/songs/musbx_api/chords_api.dart';
 import 'package:musbx/songs/musbx_api/musbx_api.dart';
-import 'package:musbx/songs/player/playable.dart';
 import 'package:musbx/songs/player/song.dart';
+import 'package:musbx/songs/player/source.dart';
 import 'package:musbx/utils/process.dart';
 
 class ChordIdentificationProcess extends Process<Map<Duration, Chord?>> {
@@ -14,14 +14,14 @@ class ChordIdentificationProcess extends Process<Map<Duration, Chord?>> {
   ChordIdentificationProcess(this.song);
 
   /// The song being analyzed.
-  final SongNew song;
+  final Song song;
 
   /// The file where the chords for this [song] are cached.
   Future<File> get cacheFile async =>
       File("${(await song.cacheDirectory).path}/chords.json");
 
   /// Perform chord analysis on the [source] using the given [host].
-  Future<Map> _analyzeSource(SongSourceNew source, ChordsApiHost host) async {
+  Future<Map> _analyzeSource(SongSource source, ChordsApiHost host) async {
     switch (source) {
       case FileSource():
         return await host.analyzeFile(source.file);
