+++ lib/songs/player/playable.dart
@@ -1,60 +1,8 @@
 import 'dart:async';
-import 'dart:io';
 
-import 'package:flutter/material.dart';
 import 'package:flutter_soloud/flutter_soloud.dart';
-import 'package:musbx/songs/demixer/demixing_process_new.dart';
 import 'package:musbx/songs/musbx_api/demixer_api.dart';
-import 'package:musbx/songs/musbx_api/musbx_api.dart';
 import 'package:musbx/songs/player/filter.dart';
-import 'package:musbx/widgets/widgets.dart';
-
-/// An object that contains information about how to [load] a [Playable].
-///
-/// This is the first step in playing a song. The [Playable] obtained by calling
-/// the [load] method can in turn be used to start playing a sound.
-abstract class SongSourceNew<P extends Playable> {
-  /// Load the [Playable] that this source points to.
-  FutureOr<P> load({required Directory cacheDirectory});
-
-  /// Free the resources used by this source.
-  FutureOr<void> dispose() {}
-
-  /// Convert this to a json map.
-  ///
-  /// The map will contain at least the following key:
-  /// - `type` [String] The type of the source.
-  ///
-  /// Depending on the type, the map will contain some additional keys. \
-  /// "youtube": `youtubeId` [String] The id of the Youtube song. \
-  /// "file": `path` [String] The path to the file. \
-  /// "demixed": `files` [Map<String, String>] The stem files.
-  Map<String, dynamic> toJson();
-
-  /// Try to create a [SongSourceNew] from a json map.
-  ///
-  /// The map should contain at least the following key:
-  /// - `type` [String] The type of the source.
-  ///
-  /// Depending on the type, the map should contain some additional keys. \
-  /// "youtube": `youtubeId` [String] The id of the Youtube song. \
-  /// "file": `path` [String] The path to the file. \
-  /// "demixed": `files` [Map<String, String>] The stem files.
-  static SongSourceNew? fromJson(Map<String, dynamic> json) {
-    if (!json.containsKey("type")) return null;
-    String? type = tryCast<String>(json["type"]);
-
-    switch (type) {
-      case "youtube":
-        return YoutubeSource.fromJson(json);
-      case "file":
-        return FileSource.fromJson(json);
-      case "demixed":
-        return DemixedSource.fromJson(json);
-    }
-    return null;
-  }
-}
 
 /// An object that can be played by [SongPlayer].
 ///
@@ -62,10 +10,10 @@ abstract class SongSourceNew<P extends Playable> {
 /// the [SongPlayer] calls when it wants to start playing the song.
 ///
 /// Note that this class is not instantiable directly but should be obtained
-/// through a [SongSourceNew], which containes instructions on how a Playable is created.
+/// through a [SongSource], which containes instructions on how a Playable is created.
 /// When you are done playing this sound you should [dispose] it to free up resources.
 ///
-/// TODO: Maybe this should be called `SongSource` and [SongSourceNew] should be `SongProvider` (or the other way around?)?
+/// TODO: Maybe this should be called `SongSource` and [SongSource] should be `SongProvider` (or the other way around?)?
 abstract class Playable {
   /// Play this sound using [SoLoud] and return the handle to the sound.
   ///
@@ -86,178 +34,9 @@ abstract class Playable {
   FutureOr<void> dispose() {}
 }
 
-class YoutubeSource extends SongSourceNew<SinglePlayable> {
-  /// A source that pulls audio from YouTube.
-  YoutubeSource(this.youtubeId);
-
-  /// The id of the YouTube song to pull.
-  final String youtubeId;
-
-  /// The [SoLoud] [AudioSource] that is generated from this source.
-  AudioSource? source;
-
-  /// The file where audio data is cached.
-  File? cacheFile;
-
-  @override
-  Future<SinglePlayable> load({required Directory cacheDirectory}) async {
-    File cacheFile = File("${cacheDirectory.path}/audio.mp3");
-
-    if (await cacheFile.exists()) {
-      debugPrint("[YOUTUBE] Using cached audio for video with id '$youtubeId'");
-    } else {
-      cacheFile = await (await MusbxApi.findYoutubeHost()).downloadYoutubeSong(
-          youtubeId,
-          destination: cacheFile,
-          fileType: "mp3");
-    }
-    this.cacheFile = cacheFile;
-
-    source ??= await SoLoud.instance.loadFile(cacheFile.path);
-
-    return SinglePlayable._(source!);
-  }
-
-  @override
-  Future<void> dispose() async {
-    if (source == null) return;
-
-    await SoLoud.instance.disposeSource(source!);
-    source = null;
-  }
-
-  /// Try to create a [YoutubeSource] from a [json] object.
-  static YoutubeSource? fromJson(Map<String, dynamic> json) {
-    if (!json.containsKey("youtubeId")) return null;
-    String? id = tryCast<String>(json["youtubeId"]);
-    if (id == null) return null;
-
-    return YoutubeSource(id);
-  }
-
-  @override
-  Map<String, dynamic> toJson() => {
-        "type": "youtube",
-        "youtubeId": youtubeId,
-      };
-}
-
-class FileSource extends SongSourceNew<SinglePlayable> {
-  /// A source that reads audio from a file.
-  FileSource(this.file);
-
-  /// The file to read.
-  final File file;
-
-  /// The [SoLoud] [AudioSource] that is generated from this source.
-  AudioSource? source;
-
-  /// The file where audio data is cached.
-  File? cacheFile;
-
-  @override
-  Future<SinglePlayable> load({required Directory cacheDirectory}) async {
-    File cacheFile = File("${cacheDirectory.path}/audio.mp3");
-
-    if (!await cacheFile.exists()) {
-      if (!await file.exists()) {
-        throw FileSystemException("File doesn't exist", file.path);
-      }
-
-      await cacheFile.create(recursive: true);
-      cacheFile = await file.copy(cacheFile.path);
-    }
-    this.cacheFile = cacheFile;
-
-    source ??= await SoLoud.instance.loadFile(cacheFile.path);
-
-    return SinglePlayable._(source!);
-  }
-
-  @override
-  Future<void> dispose() async {
-    if (source == null) return;
-
-    await SoLoud.instance.disposeSource(source!);
-    source = null;
-  }
-
-  /// Try to create a [FileSource] from a [json] object.
-  static FileSource? fromJson(Map<String, dynamic> json) {
-    if (!json.containsKey("path")) return null;
-    String? path = tryCast<String>(json["path"]);
-    if (path == null) return null;
-
-    return FileSource(File(path));
-  }
-
-  @override
-  Map<String, dynamic> toJson() => {
-        "type": "file",
-        "path": file.path,
-      };
-}
-
-class DemixedSource extends SongSourceNew<MultiPlayable> {
-  /// A source that demixes a [SongNew] and loads the stems as individual audio sources.
-  DemixedSource(this.parent);
-
-  final SongSourceNew parent;
-
-  Map<StemType, AudioSource>? sources;
-
-  @override
-  Future<MultiPlayable> load({required Directory cacheDirectory}) async {
-    DemixingProcess process = DemixingProcess(
-      parent,
-      cacheDirectory: cacheDirectory,
-    );
-
-    final Map<StemType, File> files = await process.future;
-
-    await parent.load(cacheDirectory: cacheDirectory);
-
-    sources ??= {
-      for (final e in files.entries)
-        e.key: await SoLoud.instance.loadFile(e.value.path),
-    };
-    return MultiPlayable._(sources!);
-  }
-
-  @override
-  Future<void> dispose() async {
-    if (sources == null) return;
-
-    await Future.wait([
-      for (final source in sources!.values)
-        SoLoud.instance.disposeSource(source),
-    ]);
-    sources = null;
-
-    await parent.dispose();
-  }
-
-  /// Try to create a [DemixedSource] from a [json] object.
-  static DemixedSource? fromJson(Map<String, dynamic> json) {
-    if (!json.containsKey("parent")) return null;
-    SongSourceNew? parent = SongSourceNew.fromJson(json["parent"]);
-    if (parent == null) return null;
-
-    return DemixedSource(parent);
-  }
-
-  @override
-  Map<String, dynamic> toJson() => {
-        "type": "demixed",
-        "parent": {
-          ...parent.toJson(),
-        }
-      };
-}
-
 class SinglePlayable extends Playable {
   /// A [Playable] that plays a single [source].
-  SinglePlayable._(this.source);
+  SinglePlayable(this.source);
 
   /// The source of the sound that is played.
   final AudioSource source;
@@ -284,7 +63,7 @@ class MultiPlayable extends Playable {
   /// A [Playable] that provides a voice group with a number of [sources].
   ///
   /// This allows the files to play simultaneously while the volume can be controlled individually.
-  MultiPlayable._(this.sources);
+  MultiPlayable(this.sources);
 
   /// The sources of the individual sounds that are played simultaneously.
   final Map<StemType, AudioSource> sources;
